Move implementation to a separate module and document a bunch
